+++ src/components/player.controller.js
@@ -52,6 +52,7 @@
     $scope.tagsTabDisplayed = false;
     $scope.mediaTemplate = null;
     $scope.settingsIconDisplayed = true;
+    $scope.veoLabsIconDisplayed = true;
     $scope.fullViewportActivated = false;
     $scope.templateSelectorDisplayed = true;
     $scope.fullscreenIconDisplayed = true;
@@ -1295,6 +1296,8 @@
        * @param {String} [changedProperties.oplModeIcon.currentValue] oplModeIcon new value
        * @param {Object} [changedProperties.oplSettingsIcon] oplSettingsIcon old and new value
        * @param {String} [changedProperties.oplSettingsIcon.currentValue] oplSettingsIcon new value
+       * @param {Object} [changedProperties.oplVeoLabsIcon] oplVeoLabsIcon old and new value
+       * @param {String} [changedProperties.oplVeoLabsIcon.currentValue] oplVeoLabsIcon new value
        * @param {Object} [changedProperties.oplFullViewport] oplFullViewport old and new value
        * @param {String} [changedProperties.oplFullViewport.currentValue] oplFullViewport new value
        * @param {Object} [changedProperties.oplTime] oplTime old and new value
@@ -1370,6 +1373,12 @@
           // oplSettingsIcon
           if (changedProperties.oplSettingsIcon && changedProperties.oplSettingsIcon.currentValue) {
             $scope.settingsIconDisplayed = isAttributeTrue('oplSettingsIcon', true);
+            updateIcons();
+          }
+
+          // oplVeoLabsIcon
+          if (changedProperties.oplVeoLabsIcon && changedProperties.oplVeoLabsIcon.currentValue) {
+            $scope.veoLabsIconDisplayed = isAttributeTrue('oplVeoLabsIcon', true);
             updateIcons();
           }